requirements.txt
DrissionPage>=4.0
requests
python-dotenv
apscheduler
loguru
Pillow
